shared_infra/slack_budget_bot/src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
attrs==17.4.0
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
idna==2.6                 # via requests
num2words==0.5.6
requests==2.18.4
urllib3==1.22             # via requests
